Moved tesults method and updated exception message
